Trigger InfoWindow on Autocomplete search
Autocomplete does not trigger Google's InfoWindow, thus created infoWindow and googleMarker states to do so
manually. The places_changed eventListener controls the content and visibility of the marker and info window.

Changed handleMarkerAndInfo function name to be addMarkerAndInfo because this is where the markers state is set.

Reverted back to the cancelAddMarker function existing in CreateOrEditMap because cancelling a marker entails
handling the visibility and the content of the infoWindow and googleMarker object that exists in CreateOrEditMap.
Eventually cancelAddMarker will also need to nullify other state like place.

Added logic to onMapClick method to remove the googleMarker if a user clicks a different place on the map. Did
not need to do anything with the infoWindow visibility as it automatically disappears.