Refactor async user verification call
Improved variable names in the async call that verifies new versus existing users after login. Assigned the checkForUserData return to doesUserExist in order to use its returned promise in a conditional. Previously the async functions were not in the correct sequence and resulted in errors in getting or setting user data.