Begin to add mapsSaved objects to Firestore
Set up the call to Firestore but yet to resolve an 'Unsupported field value' error. Created a testArray that uses an array of a single map object that was copied and pasted from a console.log. Succeeded in getting the testArray to update the mapsOwned field. I suspect there are stored functions, especially in the places.geometry fields, that are throwing the error.