Finish primary react-router-dom Routes
Created ViewMaps and ViewMapItinerary components to allow passing in specific mapObjects to be rendered. ViewMaps will display various maps based on the userData object, like mapsOwned, mapsSharedWithUser, and publicMapsSaved. ViewMaps sets up the Links to the ViewMapItinerary component that  uses the mapID from the url parameter to determine which map to show. Further needs here will be to combine all available maps into one array in order to determine which map is clicked by the user since it could exist in one of the three aforementioned arrays in the userData object.